# swagger.json
{
  "swagger": "2.0", 
  "info": {
    "description": "This is an API UI for DockerLogger, created by Elad Laor, for [Codefresh](https://codefresh.io)",
    "version": "1.0.0",
    "title": "DockerLogger",
    "contact": { "email": "eladlaor88@gmail.com" }
  },
  "basePath": "/",
  "tags": [
    {
      "name": "logs",
      "description": "Interact with logs on the storage layer"
    },
    {
      "name": "containers",
      "description": "View information about the containers DockerLogger is listening to"
    }
  ],
  "schemes": ["http"],
  "paths": {
    "/logs": {
      "get": {
        "tags": ["logs"],
        "summary": "Read Docker logs, and optionally filter logs",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "containerName",
            "in": "query",
            "description": "Show only logs created by a container with a matching name",
            "required": false,
            "type": "string"
          },
          {
            "name": "logType",
            "in": "query",
            "description": "Show only logs of a certain type",
            "required": false,
            "type": "string",
              "enum": ["out", "err"]
          },
          {
            "name": "minutesAgo",
            "in": "query",
            "description": "Show only messages from the past X minutes",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": { "200": { "description": "Logs retrieved" } },
        "security": []
      },
      "delete": {
        "tags": ["logs"],
        "summary": "Delete all logs messages",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": { "200": { "description": "Logs deleted" } },
        "security": []
      },
      "post": {
        "tags": ["logs"],
        "summary": "Insert an artificial log message",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": { "200": { "description": "Log entry created" } },
        "security": []
      }
    },
    "/containers": {
      "get": {
        "tags": ["containers"],
        "summary": "List the containers that are currently attached to the DockerLogger",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [      ],
        "responses": { "200": { "description": "Containers list retrieved" } },
        "security": []
      }
    },
    "/labels": {
      "post": {
        "tags": ["labels"],
        "summary": "Add a label signifier",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "labelKey",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "labelValue",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string"
          }          
        ],
        "responses": { "200": {"description": "New label inserted"} },
        "security": []
      }
    }
  },
  "definitions": {
    "Message": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "logged_message": { "type": "string" },
        "log_type": { "type": "string" },
        "container": { "type": "string" },
        "image": { "type": "string" },
        "name": { "type": "string" },
        "time": { "type": "string" }
      },
      "xml": { "name": "Message" }
    },
    "Container": {
      "type": "object",
      "properties": {
        "id": { "type": "string" },
        "name": { "type": "string" },
        "image": { "type": "string" }
      },
      "xml": { "name": "Message" }
    }
  }
}